完成物料分类部分

+++ apps/admin/src/modules/materials/mat-cate/mat-cate.controller.ts
@@ -18,20 +18,19 @@ export class MatCateController {
   @ApiOperation({ summary: "物料分类列表" })
   async index(@Response() res) {
     const data = await this.matCateService.find()
-    await this.toolsService.success(res, data)
+    await this.toolsService.success(res, this.toolsService.tree(data))
   }
 
   @Post()
   @ApiOperation({ summary: "增加物料分类"})
   async save(@Body() body:MatCateDto, @Response() res) {
     const params = body
-
     if(body.pid === 0){
       params.pid_path = "0";
       params.pid_path_name = "";
       params.level = 0
     }else {
-      const parent = await this.matCateService.find({ "pid": body.pid });
+      const parent = await this.matCateService.find({ "id": body.pid });
       if(parent[0] !== ""){
         params.pid_path = parent[0].pid_path.concat('_', `${body.pid}`);
         params.pid_path_name = parent[0].pid_path_name.concat('_', `${parent[0].cate_name}`);
@@ -49,11 +48,11 @@ export class MatCateController {
   }
 
   @Get(":id")
-  @ApiOperation({ summary: "显示一条权限详情" })
+  @ApiOperation({ summary: "显示一条分类信息" })
   async read(@Param("id") id: number, @Response() res) {
     try {
       const data = await this.matCateService.find({id: id});
-      await this.toolsService.success(res, data)
+      await this.toolsService.success(res, data[0])
     } catch (err) {
       await this.toolsService.error(res)
     }
@@ -62,8 +61,23 @@ export class MatCateController {
   @Put(":id")
   @ApiOperation({ summary: "修改权限信息"})
   async edit(@Param("id") id: number, @Body() body:MatCateDto, @Response() res) {
+    const params = body
+    if(body.pid === 0){
+      params.pid_path = "0";
+      params.pid_path_name = "";
+      params.level = 0
+    }else {
+      const parent = await this.matCateService.find({ "id": body.pid });
+      if(parent[0] !== ""){
+        params.pid_path = parent[0].pid_path.concat('_', `${body.pid}`);
+        params.pid_path_name = parent[0].pid_path_name.concat('_', `${parent[0].cate_name}`);
+        params.level = parent[0].level + 1;
+      }else {
+        await this.toolsService.error(res, "获取上级分类错误")
+      }
+    }
     try{
-      await this.matCateService.update({ "id": id }, { ...body });
+      await this.matCateService.update({ "id": id }, params);
       await this.toolsService.success(res)
     }catch (err){
       await this.toolsService.error(res, "修改错误，请重新修改", err)
@@ -72,13 +86,19 @@ export class MatCateController {
   }
 
   @Delete(":id")
-  @ApiOperation({ summary: "删除单条用户信息"})
+  @ApiOperation({ summary: "删除物料分类信息"})
   async delete(@Param("id") id: number, @Response() res) {
-    try {
-      await this.matCateService.delete({ "id": id })
-      await this.toolsService.success(res)
-    } catch (err) {
-      await this.toolsService.error(res)
+    const parent = await this.matCateService.find({pid: id})
+    if(parent.length > 0){
+      await this.toolsService.error(res, "有子分类不可删除")
+    }else {
+      try {
+        await this.matCateService.delete({ "id": id })
+        await this.toolsService.success(res)
+      } catch (err) {
+        await this.toolsService.error(res)
+      }
     }
+
   }
 }
